backend/requirements.txt
asgiref==3.11.1
Django==4.2.28
djangorestframework==3.16.1
psycopg2-binary==2.9.11
sqlparse==0.5.5
typing_extensions==4.15.0
python-dotenv==1.2.2